+++ src/main/kotlin/com/calmwolfs/bedwar/data/jsonobjects/ChatRegexJson.java
@@ -17,6 +17,7 @@ public class ChatRegexJson {
 
     @Expose
     public String selfKillPattern;
+
     @Expose
     public String teamEliminatedPattern;
 }