+++ src/main/kotlin/com/calmwolfs/bedwar/events/game/ScoreboardUpdateEvent.kt
@@ -1,0 +1,5 @@
+package com.calmwolfs.bedwar.events.game
+
+import com.calmwolfs.bedwar.events.ModEvent
+
+class ScoreboardUpdateEvent(val scoreboard: List<String>, val objective: String) : ModEvent()