+++ src/main/kotlin/com/calmwolfs/bedwar/events/game/TablistUpdateEvent.kt
@@ -1,0 +1,5 @@
+package com.calmwolfs.bedwar.events.game
+
+import com.calmwolfs.bedwar.events.ModEvent
+
+class TablistUpdateEvent(val tablist: List<String>) : ModEvent()