+++ src/main/kotlin/com/calmwolfs/bedwar/features/chat/PlayerChatClick.kt
@@ -15,12 +15,12 @@ class PlayerChatClick {
         if (event.sender == HypixelUtils.currentName) return
         if (!BedwarsUtils.inBedwarsArea) return
         if (BedWarMod.feature.chat.playerStats.clickName) {
-
-            val lastComponent =
-                if (event.component.siblings.size < 1) event.component else event.component.siblings.last()
+            val lastComponent = if (event.component.siblings.size < 1) event.component else event.component.siblings.last()
             lastComponent.chatStyle.chatClickEvent = ClickEvent(ClickEvent.Action.RUN_COMMAND, "/bws ${event.sender}")
-            lastComponent.chatStyle.chatHoverEvent =
-                HoverEvent(HoverEvent.Action.SHOW_TEXT, ChatComponentText("§aView the stats of §3${event.sender}"))
+            lastComponent.chatStyle.chatHoverEvent = HoverEvent(
+                HoverEvent.Action.SHOW_TEXT,
+                ChatComponentText("§aView the stats of §3${event.sender}")
+            )
         }
     }
 }