+++ src/main/kotlin/com/calmwolfs/bedwar/utils/HypixelUtils.kt
@@ -1,10 +1,10 @@
 package com.calmwolfs.bedwar.utils
 
 import com.calmwolfs.bedwar.BedWarMod
-import com.calmwolfs.bedwar.data.game.ScoreboardData
 import com.calmwolfs.bedwar.events.HypixelJoinEvent
-import com.calmwolfs.bedwar.events.game.ModTickEvent
+import com.calmwolfs.bedwar.events.game.ScoreboardUpdateEvent
 import net.minecraft.client.Minecraft
+import net.minecraftforge.fml.common.eventhandler.EventPriority
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import net.minecraftforge.fml.common.network.FMLNetworkEvent
 
@@ -23,32 +23,25 @@ object HypixelUtils {
     }
 
     @SubscribeEvent
-    fun onTick(event: ModTickEvent) {
-        if (!event.isMod(5)) return
-
-        if (!onHypixel) {
-            checkHypixel()
-            if (onHypixel) {
-                HypixelJoinEvent().postAndCatch()
-            }
-        }
-        if (!onHypixel) return
-
-        BedwarsUtils.bedwarsCheck()
-    }
-
-    @SubscribeEvent
     fun onHypixelJoin(event: HypixelJoinEvent) {
         BedWarMod.repo.displayRepoStatus(true)
         currentName = Minecraft.getMinecraft().thePlayer.name
     }
 
-    private fun checkHypixel() {
-        val list = ScoreboardData.scoreboard
-        if (list.isEmpty()) return
+    @SubscribeEvent(priority = EventPriority.HIGHEST)
+    fun onScoreboardUpdate(event: ScoreboardUpdateEvent) {
+        if (event.scoreboard.isEmpty()) return
 
-        val last = list.last()
-        hypixelMain = last == "§ewww.hypixel.net"
-        hypixelAlpha = last == "§ealpha.hypixel.net"
+        if (!onHypixel) {
+
+            val last = event.scoreboard.last()
+            hypixelMain = last == "§ewww.hypixel.net"
+            hypixelAlpha = last == "§ealpha.hypixel.net"
+
+            if (onHypixel) {
+                HypixelJoinEvent().postAndCatch()
+            }
+        }
+        if (!onHypixel) return
     }
 }