+++ src/main/kotlin/com/calmwolfs/bedwar/utils/PartyUtils.kt
@@ -18,11 +18,22 @@ object PartyUtils {
     @SubscribeEvent
     fun onChat(event: GameChatEvent) {
         val message = event.message.trimWhiteSpaceAndResets().removeResets()
+        val unformatted = message.unformat()
+
+        // no word party
+        "§eKicked (?<name>.*) because they were offline.".toPattern().matchMatcher(message) {
+            val name = group("name").toPlayerName()
+            partyMembers.remove(name)
+            return
+        }
+
+        if (!unformatted.lowercase().contains("party")) return
 
         // new member joined
         "§eYou have joined (?<name>.*)'s §eparty!".toPattern().matchMatcher(message) {
             val name = group("name").toPlayerName()
             if (!partyMembers.contains(name)) partyMembers.add(name)
+            return
         }
         "(?<name>.*) §ejoined the party.".toPattern().matchMatcher(message) {
             val name = group("name").toPlayerName()
@@ -30,61 +41,68 @@ object PartyUtils {
                 partyMembers.add(name)
                 PlayerJoinPartyEvent(name).postAndCatch()
             }
+            return
         }
         "§eYou'll be partying with: (?<names>.*)".toPattern().matchMatcher(message) {
             for (name in group("names").split(", ")) {
                 val playerName = name.toPlayerName()
                 if (!partyMembers.contains(playerName)) partyMembers.add(playerName)
             }
+            return
         }
 
         // one member got removed
         "(?<name>.*) §ehas left the party.".toPattern().matchMatcher(message) {
             val name = group("name").toPlayerName()
             partyMembers.remove(name)
+            return
         }
         "(?<name>.*) §ehas been removed from the party.".toPattern().matchMatcher(message) {
             val name = group("name").toPlayerName()
             partyMembers.remove(name)
-        }
-        "§eKicked (?<name>.*) because they were offline.".toPattern().matchMatcher(message) {
-            val name = group("name").toPlayerName()
-            partyMembers.remove(name)
+            return
         }
         "(?<name>.*) §ewas removed from your party because they disconnected.".toPattern().matchMatcher(message) {
             val name = group("name").toPlayerName()
             partyMembers.remove(name)
+            return
         }
-        "The party was transferred to .* because (?<name>.*) left".toPattern().matchMatcher(message.unformat()) {
+        "The party was transferred to .* because (?<name>.*) left".toPattern().matchMatcher(unformatted) {
             val name = group("name").toPlayerName()
             partyMembers.remove(name)
+            return
         }
 
         // party disbanded
         ".* §ehas disbanded the party!".toPattern().matchMatcher(message) {
             partyMembers.clear()
+            return
         }
         "§eYou have been kicked from the party by .* §e".toPattern().matchMatcher(message) {
             partyMembers.clear()
+            return
         }
         if (message == "§eYou left the party." ||
             message == "§cThe party was disbanded because all invites expired and the party was empty." ||
             message == "§cYou are not currently in a party."
         ) {
             partyMembers.clear()
+            return
         }
 
         // party list
         "§6Party Members \\(\\d+\\)".toPattern().matchMatcher(message) {
             partyMembers.clear()
+            return
         }
 
-        "Party (?:Leader|Moderators|Members): (?<names>.*)".toPattern().matchMatcher(message.unformat()) {
+        "Party (?:Leader|Moderators|Members): (?<names>.*)".toPattern().matchMatcher(unformatted) {
             for (name in group("names").split(" ● ")) {
                 val playerName = name.replace(" ●", "").toPlayerName()
                 if (playerName == Minecraft.getMinecraft().thePlayer.name) continue
                 if (!partyMembers.contains(playerName)) partyMembers.add(playerName)
             }
+            return
         }
     }